[SWM-185] ERD 리팩토링 반영

+++ src/main/java/com/climbx/climbx/problem/dto/ProblemDetailsResponseDto.java
@@ -11,7 +11,7 @@ public record ProblemDetailsResponseDto(
     String gymName,
     String localLevel,
     String holdColor,
-    Long problemRating,
+    Integer problemRating,
     Long spotId,
     Double spotXRatio,
     Double spotYRatio,
